feat: Implement adding team member and owner info to Project (#31)

+++ src/project/service.py
@@ -1,26 +1,40 @@
-from typing import List
-from fastapi import UploadFile
+from typing import List, Optional
+from fastapi import File, UploadFile
 from sqlalchemy.orm import Session
 
-from src.models import Project
-from project.schemas import ProjectReq, ProjectRes
+from src.models import Project, ProjectUser
+from project.schemas import ProjectReq, ProjectRes, ProjectUserRes
 from project import repository
-from src.exceptions.definitions import ProjectAlreadyExist, ProjectNotFound
+from src.exceptions.definitions import (
+    ProjectAlreadyExist,
+    ProjectNotFound,
+    ProjectOwnerMismatched,
+)
+
+from src.project_user.repository import create_project_user
+from src.user.repository import find_user_by_user_id
+from src.user.schemas import UserRes
 
 
-def create_project(project_req: ProjectReq, files: List[UploadFile], db: Session):
+def create_project(
+    user_id: int,
+    project_req: ProjectReq,
+    db: Session,
+    files: Optional[List[UploadFile]] = File(None),
+):
     """
     Create a new project
 
     Returns project data
     """
-    existing_project = repository.find_project_by_name(db, project_req.project_name)
+    existing_project = repository.find_project_by_name(db, project_req.name)
 
     if existing_project:
         raise ProjectAlreadyExist()
 
     project = Project(
-        project_name=project_req.project_name,
+        name=project_req.name,
+        owner=user_id,
         repo_name=project_req.repo_name,
         start_date=project_req.start_date,
         end_date=project_req.end_date,
@@ -29,9 +43,22 @@ def create_project(project_req: ProjectReq, files: List[UploadFile], db: Session
     )
 
     saved_project = repository.create_project(db, project)
-    project_res = ProjectRes.model_validate(saved_project)
+    project_members = []
+
+    for req_member in project_req.members:
+        found_user = find_user_by_user_id(db, req_member.id)
+        project_user = ProjectUser(
+            user=found_user, project=saved_project, role=req_member.role
+        )
+        saved_project_user = create_project_user(db, project_user)
+        project_member = ProjectUserRes.from_user(found_user, saved_project_user.role)
+        project_members.append(project_member)
+
+    owner_user = find_user_by_user_id(db, user_id)
+    project_res = ProjectRes.from_project(saved_project, owner_user, project_members)
 
     # TODO Embedding project files
+    # Check files exist
     # saved_files = await upload_file(saved_project.project_name, files)
 
     return project_res
@@ -47,7 +74,15 @@ def get_project(project_id: int, db: Session):
     if not existing_project:
         raise ProjectNotFound()
 
-    project_res = ProjectRes.model_validate(existing_project)
+    owner_user = find_user_by_user_id(db, existing_project.owner)
+
+    project_members = []
+    for project_user in existing_project.members:
+        found_user = find_user_by_user_id(db, project_user.user_id)
+        project_member = ProjectUserRes.from_user(found_user, project_user.role)
+        project_members.append(project_member)
+
+    project_res = ProjectRes.from_project(existing_project, owner_user, project_members)
 
     return project_res
 
@@ -64,7 +99,7 @@ def update_project(
     if not existing_project:
         raise ProjectNotFound()
 
-    existing_project.project_name = (project_req.project_name,)
+    existing_project.name = (project_req.name,)
     existing_project.repo_name = (project_req.repo_name,)
     existing_project.start_date = (project_req.start_date,)
     existing_project.end_date = (project_req.end_date,)
@@ -72,15 +107,24 @@ def update_project(
     existing_project.discord_channel_id = (project_req.discord_chnnel_id,)
 
     saved_project = repository.update_project(db, existing_project)
-    project_res = ProjectRes.model_validate(saved_project)
 
     # TODO Embedding project files
     # saved_files = await upload_file(saved_project.project_name, files)
 
+    owner_user = find_user_by_user_id(db, saved_project.owner)
+
+    project_members = []
+    for project_user in saved_project.members:
+        found_user = find_user_by_user_id(db, project_user.user_id)
+        project_member = ProjectUserRes.from_user(found_user, project_user.role)
+        project_members.append(project_member)
+
+    project_res = ProjectRes.from_project(saved_project, owner_user, project_members)
+
     return project_res
 
 
-def delete_project(project_id: int, db: Session):
+def delete_project(user_id: int, project_id: int, db: Session):
     """
     Delete the existing project by project id
     """
@@ -88,7 +132,10 @@ def delete_project(project_id: int, db: Session):
     if not existing_project:
         raise ProjectNotFound()
 
-    repository.delete_project(db, existing_project)
+    if existing_project.owner == int(user_id):
+        repository.delete_project(db, existing_project)
+    else:
+        raise ProjectOwnerMismatched()
 
 
 async def upload_file(project_name: str, files: List[UploadFile]):
